+++ linux/shell/wp/remove-comps.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -xe
 
 # author rajen
 # date 1/19/2017
@@ -20,6 +21,7 @@ DB_PASSWD=$1
 DB_WORDPRESS_NAME='wordpress'
 DB_WORDPRESS_USER=$1
 DB_WORDPRESS_PASSWD=$1
+PHP_VERSION=$2
 
 # remove LAMP
 #   * Linux
@@ -38,15 +40,17 @@ function apache_remove {
     sudo apt-get remove -y apache2
 	sudo mv /etc/apache2/apache2.conf.bk /etc/apache2/apache2.conf
 	sudo mv /etc/hosts.bk /etc/hosts
-	sudo service apache2 restart
 }
 
 ## remove MySQL and create database:wordpress 
 function mysql_remove {
-    sudo apt-get remove -y mysql-server php5-mysql
-	sudo mv /etc/mysql/my.cnf.bk /etc/mysql/my.cnf 
-	sudo service mysql restart
-
+    sudo apt-get remove -y mysql-server php${PHP_VERSION}-mysql
+    sudo apt-get remove -y mysql-common mysql-client
+    if [[ -f /etc/mysql/my.cnf.bk ]]; then
+        sudo mv /etc/mysql/my.cnf.bk /etc/mysql/my.cnf
+    else
+        sudo cp /etc/mysql/my.cnf /etc/mysql/my.cnf.og
+    fi
 }
 
 
@@ -59,14 +63,21 @@ function mysql_clean_db {
 
 ## remove php
 function php_remove {
-    sudo apt-get remove -y php5 libapache2-mod-php5 php5-mcrypt
-    sudo apt-get remove -y php5-gd php5-curl libssh2-php
-    sudo apt-get remove -y php5-cli
-    sudo apt-get remove -y php5-mysqlnd-ms
+    sudo apt-get remove -y php${PHP_VERSION} libapache2-mod-php${PHP_VERSION} php${PHP_VERSION}-mcrypt
+    sudo apt-get remove -y php${PHP_VERSION}-gd php${PHP_VERSION}-curl libssh2-php
+    sudo apt-get remove -y php${PHP_VERSION}-cli
+    sudo apt-get remove -y php${PHP_VERSION}-mysqlnd-ms
 }
 
-function php_config_remove {
-	sudo mv /etc/php5/apache2/php.ini.bk /etc/php5/apache5/php.ini
+function php5_config_remove {
+	sudo mv /etc/php5/apache2/php.ini.bk /etc/php5/apache2/php.ini
+	sudo mv /etc/apache2/mods-enabled/dir.conf.bk /etc/apache2/mods-enabled/dir.conf
+    sudo service apache2 restart
+    sudo a2enmod rewrite
+    
+}
+function php5X_config_remove {
+	sudo mv /etc/php/${PHP_VERSION}/apache2/php.ini.bk /etc/php/${PHP_VERSION}/apache2/php.ini
 	sudo mv /etc/apache2/mods-enabled/dir.conf.bk /etc/apache2/mods-enabled/dir.conf
     sudo service apache2 restart
     sudo a2enmod rewrite
@@ -74,18 +85,18 @@ function php_config_remove {
 }
 
 
+
 # download wordpress
 function wordpress_content_remove {
     cd /var/www/html
 	sudo rm -rf *
-    sudo service apache2 restart
 }
 
-end_remove
-apache_remove
-mysql_clean_db
-mysql_remove
-php_remove
-php_config_remove
+#end_remove
+#mysql_clean_db
+#mysql_remove
+#php5X_config_remove
+#php5_remove
+#apache_remove
 wordpress_content_remove
-
+end_remove
